// packages/core/src/types/room/floor-layout.ts
/**
 * Тип раскладки напольного покрытия.
 *
 * @public
 * @enum
 */
export enum FloorLayout {
  /**
   * Прямая палубная раскладка.
   *
   * @public
   * @member
   */
  Straight = 'straight',

  /**
   * Раскладка ёлочкой.
   *
   * @public
   * @member
   */
  Herringbone = 'herringbone',
}